- fixed bug 4791 (http://bugzilla.osafoundation.org/show_bug.cgi?id=4791)
git-svn-id: http://svn.osafoundation.org/chandler/trunk@8485 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/repository/persistence/DBRepository.py
@@ -291,7 +291,8 @@ class DBRepository(OnDemandRepository):
             restore = kwds.get('restore', None)
 
             if restore is not None and os.path.isdir(restore):
-                self.delete()
+                if os.path.exists(self.dbHome):
+                    self.delete()
                 if not os.path.exists(self.dbHome):
                     os.mkdir(self.dbHome)
                 for f in os.listdir(restore):
